Password Generator - [Task 3]
This program is a Python password generator. It uses the `random` and `string` modules to generate a random password of a specified length. The password can contain any combination of letters, numbers, and punctuation symbols.

Here is a more detailed description of the program:

* The `generate_password()` function takes the desired length of the password as input and returns a randomly generated password of that length.
* The `main()` function prompts the user to enter the desired length of the password and then calls the `generate_password()` function to generate the password.
* The `main()` function then prints the generated password to the console.

This program is a good example of how to use the `random` and `string` modules to generate random strings. It is also a good example of how to write a simple Python program to perform a specific task.

Here is an example of how to use the program:

```
Enter the desired length of the password: 12
Generated Password: Wl?XaSu-JsD(
```